ajustes routes

+++ my-app-laravel/routes/web.php
@@ -10,13 +10,21 @@ Route::get('/', function () {
     return view('home');
 });
 
-Route::get('/users', [UserController::class, 'index'])->name('users.index');
-Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update');
+Route::middleware(['auth'])->group(function (){
 
-Route::get('/products', [ProductController::class, 'index'])->name('products.index');
-Route::get('/products/create', [ProductController::class, 'create'])->name('products.create');
-Route::post('/product', [ProductController::class, 'store'])->name('products.store');
-Route::delete('/products/{id}', [ProductController::class, 'destroy'])->name('products.destroy');
-Route::get('/products/{id}/edit', [ProductController::class, 'edit'])->name('products.edit');
-Route::put('/products/{id}', [ProductController::class, 'update'])->name('products.update');
+    Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit')->middleware('auth');
+    Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update')->middleware('auth');
+    Route::get('/products', [ProductController::class, 'index'])->name('products.index')->middleware('auth');
+
+});
+
+Route::middleware(['auth','admin'])->group(function (){
+
+    Route::get('/users', [UserController::class, 'index'])->name('users.index')->middleware('admin');
+    Route::get('/products/create', [ProductController::class, 'create'])->name('products.create')->middleware('admin');
+    Route::post('/product', [ProductController::class, 'store'])->name('products.store')->middleware('admin');
+    Route::delete('/products/{id}', [ProductController::class, 'destroy'])->name('products.destroy')->middleware('admin');
+    Route::get('/products/{id}/edit', [ProductController::class, 'edit'])->name('products.edit')->middleware('admin');
+    Route::put('/products/{id}', [ProductController::class, 'update'])->name('products.update')->middleware('admin');
+
+});
